Fixed imports for chromagnonCache; implemented pathlib
I fixed the imports required for the chromagnonCache script. I was able to run this script successfully.
I implemented "pathlib" to solve for difference between OS. The original script wasn't tested in Windows; but with use of Path from pathlib this issue should be mitigated. I updated all places where I hit an error due to the filepath. Future testing will be done to ensure it works properly on MacOS/Linux.