Add support for additional SNSS commands.
Additional support has been added for some SNSS commands. However, I've commented one out due to what appears to be lack of forensic significance. At a future point, I plan to have a default subset for commands to return and an optional flag to specify for all of the SNSS commands to be parsed.